+++ app.py
@@ -4,7 +4,7 @@ from sys import exit
 from sptfy import Sptfy
 from pathlib import Path
 from commands import get_usage, top_level_command_args
-from helpers import get_required_environment_variables, ENVIRONMENT_ENV_VARIABLE_STR
+from helpers import get_required_environment_variables
 from commands import configure
 
 
@@ -23,11 +23,11 @@ def main() -> None:
         configure.configure_command()
         exit(0)
 
-    spotify_client_id, spotify_client_secret, spotify_redirect_url = get_required_environment_variables()
+    env_vars = get_required_environment_variables()
     sptfy = Sptfy(
-        spotify_client_id=spotify_client_id,
-        spotify_client_secret=spotify_client_secret,
-        spotify_redirect_uri=spotify_redirect_url,
+        spotify_client_id=env_vars.spotify_client_id,
+        spotify_client_secret=env_vars.spotify_client_secret,
+        spotify_redirect_uri=env_vars.spotify_redirect_uri,
     )
 
     command_function = top_level_command_args.get(command, None)
@@ -42,9 +42,10 @@ def main() -> None:
 
 
 if __name__ == "__main__":
-    environment = os.environ.get(f"{ENVIRONMENT_ENV_VARIABLE_STR}", "development")
-
-    if environment == "production":
+    # Get parent dir and switch the working directory
+    # When run as an executable, the working directory is a temp folder,
+    # using this we can get the folder of the actual file
+    if getattr(sys, 'frozen', False):
         executable_path = Path(sys.executable)
         parent_dir = executable_path.parent.absolute()
     else: