+++ sptfy/sptfy.py
@@ -116,8 +116,9 @@ class Sptfy:
             auth_manager=auth_manager
         )
 
-    def auth(self) -> None:
+    def auth(self) -> bool:
         self.spotify.current_user()
+        return True
 
     def get_all_playlists_no_songs(self, limit=50, offset=0) -> List[PlaylistNoSongs]:
         logger.info(f"Retrieving all playlists")